+++ cmd/serve/serve.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	e "errors"
 	"fmt"
-	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/usecases/relationships"
-	resourcesctl "github.com/project-kessel/inventory-api/internal/biz/usecases/resources"
+	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/usecase/relationships"
+	resourcesctl "github.com/project-kessel/inventory-api/internal/biz/usecase/resources"
 	"os"
 	"os/signal"
 	"syscall"
