+++ internal/service/relationships/k8spolicy/ispropagatedto_k8scluster.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
 	"github.com/project-kessel/inventory-api/internal/biz/model"
-	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/usecases/relationships"
+	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/usecase/relationships"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	conv "github.com/project-kessel/inventory-api/internal/service/common"
